cleanup docs and options
