+++ CHANGELOG.md
@@ -15,17 +15,20 @@ All notable changes to the ld-find-code-refs program will be documented in this 
     defaultsDisabled: true
   ```
 - All command line flags can now be specified as environment variables.
-- A new option `searchTool` or `LD_SEARCH_TOOL` (as an environment variable) to specify the underlying tool used for search. Ripgrep (rg) and Silver Searcher (ag) are currently supported.
+
+### Fixed
+
+- Exlude negations in `.ldignore` (lines beginning with an exlamation mark) now correctly include files.
 
 ### Changed
 
-- The underlying command line package has been changed for improved ergonomics when configuring options using the command line. Some existing configurations may be invalid, see `ld-find-code-refs --help` for details.
-- Ripgrep (rg) is now the default search tool. Ripgrep will need to be installed in the system path, or the `searchTool` option set to `ag` for `ld-find-code-refs` to continue to work.
+- The command line tool has been changed for improved ergonomics when configuring options using the command line. Some existing configurations may be invalid, see `ld-find-code-refs --help` for details.
 
 ### Removed
 
 - The `exclude` command-line option has been removed. Use the `.ldignore` file instead.
 - Additional delimiters can no longer be specified using command line flags or environment variables. If you use additional delimiters, or would like to disable delimiters completely, use the YAML configuration described above.
+- `ld_find_code_refs` no longer requires the silver searcher (ag) as a runtime dependency.
 
 ## [1.5.0] - 2020-05-11
